refactor(web): reorganize section cards into layout cards + small tweaks
