+++ apps/web/src/components/Cards/Bases.tsx
@@ -9,11 +9,13 @@ export const CardTitleBase: FC<CardTitleBaseProps> = function ({
   children,
   type = "header",
 }) {
+  const isHeader = type === "header";
   return (
     <div
       className={`
-        -mx-4
-        ${type === "header" ? "-mt-4" : "-mb-4"}
+        ${isHeader ? "rounded-t-md" : "rounded-b-md"}
+        ${isHeader ? "-mx-4" : "-mx-6"}
+        ${type === "header" ? "-mt-4" : "-mb-6"}
         bg-surfaceHeader-light
         p-3
         text-base
