+++ apps/web/src/components/Header.tsx
@@ -1,0 +1,5 @@
+import { type FC, type ReactNode } from "react";
+
+export const Header: FC<{ children: ReactNode }> = function ({ children }) {
+  return <div className="truncate text-2xl font-bold">{children}</div>;
+};
