+++ apps/web/src/components/InfoGrid.tsx
@@ -1,6 +1,6 @@
 import React, { Fragment, type ReactNode } from "react";
 
-type InfoGridProps = {
+export type InfoGridProps = {
   fields: { name: ReactNode; value: ReactNode }[];
 };
 
@@ -9,7 +9,7 @@ export const InfoGrid: React.FC<InfoGridProps> = function ({ fields }) {
     <div className="grid w-fit gap-3 md:grid-cols-4">
       {fields.map(({ name, value }, i) => (
         <Fragment key={i}>
-          <div className="font-bold dark:text-coolGray-400">{name}</div>
+          <div className="font-semibold dark:text-coolGray-400">{name}</div>
           <div className="col-span-3 truncate text-sm">{value}</div>
         </Fragment>
       ))}
