+++ apps/web/src/utils/date.ts
@@ -1,7 +1,8 @@
 import dayjs from "./dayjs";
 
-export function formatTimestamp(timestamp: number) {
-  const unixHandler = dayjs.unix(timestamp);
+export function formatTimestamp(timestamp: number | Date) {
+  const unixHandler =
+    typeof timestamp === "number" ? dayjs.unix(timestamp) : dayjs(timestamp);
 
   return `${unixHandler.fromNow()} (${unixHandler.format(
     "MMM D, YYYY h:mm AZ",
